requirements.txt
reportlab>=4
pillow>=10
streamlit==1.35.0
pyarrow==16.1.0